requirements.txt
tensorflow
transformers
tensorflow_addons
numpy
pandas
pydot
scikit-learn
onnx==1.8.1
pytorch2keras
scispacy
spacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.0/en_core_sci_lg-0.5.0.tar.gz
biopython